requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
PyMuPDF==1.23.8
pdfplumber==0.10.3
opencv-python-headless==4.8.1.78
python-dotenv==1.0.0
pydantic>=2.7.0
pydantic-settings
numpy<2